requirements.txt
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cors==1.0.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pillow==12.0.0
pycparser==2.23
PySocks==1.7.1
requests==2.32.5
requests-file==3.0.1
sympy==1.14.0
tldextract==5.3.0
torch==2.9.0
torchvision==0.24.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0.1
